taskmanager/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.25.0
kafka-python-ng==2.2.0
# python-rocksdb==0.7.0  # Commented out - using in-memory backend
prometheus-client==0.19.0
boto3==1.29.0
